% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello world}
\usage{
hello(name = "Jamie")
}
\arguments{
\item{name}{(character; default = "Jamie") Name of person to greet}
}
\value{
Nothing
}
\description{
Hello world
}
\examples{
\dontrun{
hello("Jamie")
}
}
